Fix racecar speed issue and quality of life improvements.
Racecar:
- Fix speed discrepancy between different framerates.

General:
- Player can now exit to the track selection screen before exiting the game.
- General improvements to code.

AI:
- Improve collision
- Remove rubberbanding
- Add random unique speed for each instance

Nodes:
- Fix bug where radius was stored as an int

Racecar:
- Add damage time buffer to avoid damage chaining

Racetrack:
- Move Control keybinding to parameter list
- Improve scenery radiuses

RNG:
- Create RNG module to generate random numbers

Startup Screen:
- Remove accelerating animation
- Shorten time

+++ node.h
@@ -30,7 +30,7 @@ namespace desert
 	class SceneNodeContainer
 	{
 	public:
-		static NodeAlignment getAlignmentFromRotation(int rotation);
+		static NodeAlignment getAlignmentFromRotation(float rotation);
 		/**
 		* @param sceneNode A TL-Engine ISceneNode pointer
 		*/
@@ -116,7 +116,7 @@ namespace desert
 		virtual Collision::CollisionAxis collision(SphereCollisionModel other);
 		virtual int getCollisionRadius();
 	protected:
-		int mRadius;
+		float mRadius;
 	};
 
 	class BoxCollisionModel : public CollisionModel
